requirements.txt
streamlit
openai
Pillow
requests
python-dotenv
jupytext
streamlit-webrtc
gradio 
openai 
torch

